requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
catkin-pkg==1.0.0
fvcore==0.1.5.post20221221
gltflib==1.0.13
h5py==3.8.0
hello-robot-stretch-body==0.7.12
hello-robot-stretch-body-tools==0.7.7
hello-robot-stretch-factory==0.5.5
hello-robot-stretch-tool-share==0.3.4
hello-robot-stretch-urdf==0.0.23
importlib-metadata==7.0.0
matplotlib==3.5.0
matplotlib-inline==0.1.6
numpy==1.23.2
numpy-quaternion==2022.4.2
numpy-stl==3.1.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
open3d==0.17.0
opencv-contrib-python==4.8.1.78
opencv-python==4.7.0.68
pathlib==1.0.1
pillow==9.4.0
pin==2.6.17
pycocotools==2.0.6
pyrealsense2
rospkg==1.5.0
scikit-image
scikit-learn==1.3.2
scipy==1.10.1
tabulate==0.9.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
